git-svn-id: https://svn.vis.ethz.ch/svn/kob_gamelab09/trunk@1005 fc16ba98-2866-4055-be8f-df951a5456ab

+++ src/ProjectMagma/ProjectMagma/Simulation/Properties/Player/PlayerControllerProperty.cs
@@ -1129,7 +1129,7 @@ namespace ProjectMagma.Simulation
                 normal.Normalize();
 
             // and hit?
-            if (simTime.At < controllerInput.hitButtonPressedAt + constants.GetFloat("hit_cooldown")
+            if (simTime.At < controllerInput.hitButtonPressedAt + constants.GetInt("hit_cooldown")
                 && simTime.At > hitPerformedAt + constants.GetInt("hit_cooldown"))
             {
                 // indicate hit!
